v2.7.4: merge settings and config close #6
