+++ v2/src/codeRunner/brythonRunner.ts
@@ -1,4 +1,3 @@
-import Result from "./Result";
 import CustomRunner from "./CustomRunner";
 
 let brythonRunnerLoaded = false;
