+++ v2/src/config/page.html
@@ -9,7 +9,7 @@
   <body>
     <div class="container">
       <div class="panel panel-default">
-        <div class="panel-heading">Settings</div>
+        <div class="panel-heading">config</div>
         <div class="panel-body">
           <form id="options">
           </form>
