+++ v2/src/site/about.ts
@@ -1,6 +1,6 @@
 import TestCase from "../TestCase";
 import { loadScript, newElement, ObservableValue } from "../util";
-import settings from "../settings";
+import config from "../config";
 
 async function init() {
   if (location.host != "greasyfork.org" && !location.href.match(/433152-atcoder-easy-test-v2/)) throw "Not about page";
@@ -15,8 +15,8 @@ async function init() {
 
   doc.getElementById("install-area").appendChild(newElement("button", {
     type: "button",
-    textContent: "Open Settings",
-    onclick: () => settings.open(),
+    textContent: "Open config",
+    onclick: () => config.open(),
   }));
 
   return {
